AI Project/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dateutil>=2.8.2
gunicorn>=21.2.0
lxml>=5.0.0
googlenewsdecoder>=0.1.7
